+++ src/index.ts
@@ -38,7 +38,7 @@ import { BotConfig } from './typeorm/entities/BotConfig';
 import { BaitChannelConfig } from './typeorm/entities/bait/BaitChannelConfig';
 import { BaitChannelLog } from './typeorm/entities/bait/BaitChannelLog';
 import { BaitKeyword } from './typeorm/entities/bait/BaitKeyword';
-import { PendingBan } from './typeorm/entities/PendingBan';
+import { PendingBan } from './typeorm/entities/bait/PendingBan';
 import { UserActivity } from './typeorm/entities/UserActivity';
 import type { ExtendedClient } from './types/ExtendedClient';
 import {